+++ libraries/ui_compose_material3/src/main/java/androidx/media3/ui/compose/material3/MiniControllerTokens.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2026 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.media3.ui.compose.material3
+
+import androidx.compose.ui.unit.dp
+
+internal object MiniControllerTokens {
+  val ArtworkSize = 56.dp
+  val CardElevation = 4.dp
+  val ControlSize = 48.dp
+  val HorizontalPadding = 8.dp
+  val VerticalPadding = 4.dp
+  val SpacerWidth = 8.dp
+}